services/lab-service/requirements.txt
flask==3.0.0
prometheus-flask-exporter==0.23.0
psycopg2-binary==2.9.9
pytest==7.4.0
pytest-flask==1.3.0
